Creates new test. Moves factories over to test from reporting_base
