+++ modules/claims_api/app/sidekiq/claims_api/reporting_base.rb
@@ -154,64 +154,5 @@ module ClaimsApi
       # Group the array by the new keys
       array.group_by { |item| transaction_mapping[item[:transaction_id]] }
     end
-
-    def call_factories
-      make_claims
-      make_poas
-      make_ews_submissions
-      make_poas
-    end
-
-    def make_claims
-      FactoryBot.create(:auto_established_claim_v2, status: 'errored')
-      FactoryBot.create(:auto_established_claim_v2, status: 'errored')
-
-      FactoryBot.create(:auto_established_claim_va_gov, created_at: Time.zone.now)
-      FactoryBot.create(:auto_established_claim_va_gov, created_at: Time.zone.now)
-      FactoryBot.create(:auto_established_claim_va_gov, :transaction_id_25, created_at: Time.zone.now)
-      FactoryBot.create(:auto_established_claim_va_gov, :transaction_id_25, created_at: Time.zone.now)
-
-      FactoryBot.create(:auto_established_claim_v2, status: 'errored')
-      FactoryBot.create(:auto_established_claim_v2, status: 'pending')
-      FactoryBot.create(:auto_established_claim_without_flashes_or_special_issues)
-      FactoryBot.create(:auto_established_claim_without_flashes_or_special_issues)
-      FactoryBot.create(:auto_established_claim_with_supporting_documents)
-      FactoryBot.create(:auto_established_claim)
-    end
-
-    def make_poas
-      FactoryBot.create(:power_of_attorney, :errored)
-      FactoryBot.create(:power_of_attorney, :errored)
-      FactoryBot.create(:power_of_attorney)
-      FactoryBot.create(:power_of_attorney)
-    end
-
-    def make_ews_submissions
-      FactoryBot.create(:claims_api_evidence_waiver_submission, :errored)
-      FactoryBot.create(:claims_api_evidence_waiver_submission)
-      FactoryBot.create(:claims_api_evidence_waiver_submission, :errored)
-      FactoryBot.create(:claims_api_evidence_waiver_submission)
-    end
-
-    def make_itfs
-      FactoryBot.create(:claims_api_intent_to_file, :itf_errored)
-      FactoryBot.create(:claims_api_intent_to_file, :itf_errored)
-      FactoryBot.create(:claims_api_intent_to_file)
-    end
-
-    def gather_consumers
-      @claims_consumers = ClaimsApi::AutoEstablishedClaim.where(created_at: @from..@to).pluck(:cid).uniq
-      @poa_consumers = ClaimsApi::PowerOfAttorney.where(created_at: @from..@to).pluck(:cid).uniq
-      @ews_consumers = ClaimsApi::EvidenceWaiverSubmission.where(created_at: @from..@to).pluck(:cid).uniq
-      @itf_consumers = ClaimsApi::IntentToFile.where(created_at: @from..@to).pluck(:cid).uniq
-    end
-
-    def db_clean
-      # destroys anything created in the last 24 hours
-      ClaimsApi::AutoEstablishedClaim.where(created_at: @from..@to).destroy_all
-      ClaimsApi::PowerOfAttorney.where(created_at: @from..@to).destroy_all
-      ClaimsApi::EvidenceWaiverSubmission.where(created_at: @from..@to).destroy_all
-      ClaimsApi::IntentToFile.where(created_at: @from..@to).destroy_all
-    end
   end
 end
