+++ modules/claims_api/app/views/claims_api/unsuccessful_report_mailer/unsuccessful_report.html.erb
@@ -66,12 +66,6 @@
     </h3>
     <%= render partial: 'claims_api/submission_grouped_table', locals: { claims: @unsuccessful_va_gov_claims_submissions } unless @unsuccessful_va_gov_claims_submissions&.count&.zero? %>
 
-    <h3>
-      <%= @unsuccessful_va_gov_claims_submissions&.count%>
-      <span class="title">526 VA GOV Errored Submissions</span>
-    </h3>
-    <%= render partial: 'claims_api/submission_grouped_table', locals: { claims: @unsuccessful_va_gov_claims_submissions } unless @unsuccessful_va_gov_claims_submissions&.count&.zero? %>
-    
     <hr>
     <h2>Power of Attorney Submissions</h2>
